+++ examples/simple/index.js
@@ -12,11 +12,25 @@ const logger = new Logger({
 });
 
 setInterval(() => {
-  logger.info("Hey Buddah!", { pid: process.pid });
+  logger.info("Hey Buddah!", {
+    pid: process.pid,
+    env: process.env.NODE_ENV,
+    ...process.memoryUsage(),
+  });
 }, 1000);
+
+setInterval(() => {
+  logger.warn("Monkey, watch out!", {
+    pid: process.pid,
+    env: process.env.NODE_ENV,
+    ...process.memoryUsage(),
+  });
+}, 7000);
 
 setInterval(() => {
   logger.error("I love a good fight!", new Error("Oooh, Demons!"), {
     pid: process.pid,
+    env: process.env.NODE_ENV,
+    ...process.memoryUsage(),
   });
 }, 3000);