+++ test/processors/human.test.js
@@ -6,7 +6,7 @@ const {
 
 describe("human", () => {
   it("should work out of the box", () => {
-    const fn = human();
+    const fn = human({ colours: 0 });
     const result = fn({
       record: {
         level: Level.INFO.name,
@@ -15,12 +15,12 @@ describe("human", () => {
     });
     match(
       result,
-      /^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} \[INFO\] How blissful it is, for one who has nothing$/
+      /^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} INFO  How blissful it is, for one who has nothing$/
     );
   });
 
   it("should work with errors out of the box", () => {
-    const fn = human();
+    const fn = human({ colours: 0 });
     const error = new Error("Oooh, Demons!");
     const result = fn({
       record: {
@@ -29,10 +29,7 @@ describe("human", () => {
         error: { message: error.message, stack: error.stack },
       },
     });
-    match(
-      result,
-      /^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} \[ERROR\] I forbid it!/
-    );
+    match(result, /^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} ERROR I forbid it!/);
     match(result, /Error: Oooh, Demons!/);
   });
 });